refactor: 체크리스트 질문 answer 관련 훅으로 리팩토링
